backend/lambda-email-minimal/lambda-csv-minimal/requirements.txt
boto3>=1.34.0
PyJWT>=2.8.0
bcrypt>=4.0.1
fastapi>=0.104.0
mangum>=0.17.0
Pillow>=10.0.0
imagehash>=4.3.1
requests>=2.31.0



